feat: MemoryStorage 삭제 및 RealmStorage 구현
