+++ ProjectManager/ProjectManager/Application/DIContainer/AppDIContainer.swift
@@ -12,7 +12,7 @@ protocol AppDIContainerable {
 }
 
 final class AppDIContainer: AppDIContainerable {
-    private let storage = MemoryStorage()
+    private let storage = RealmStorage()
     
     func makeTodoListSceneDIContainer() -> TodoListSceneDIContainer {
         TodoListSceneDIContainer(dependencies: TodoListSceneDIContainer.Dependencies(storage: storage))
